cmd/pyproject.toml
[project]
name = "vertretungsplan-cmd"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
